# gestiondestock_frontend/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger Api documentation for Gestion de stock application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger Api Documentation for Gestion de stock application",
    "contact" : {
      "name" : "Wissem Belhassen",
   
      "email" : "belhassenwissem@gmail.com"
    }
  },
  "basePath" : "http://localhost:8081",
...
  "paths" : {
...
    "/gestiondestock/v1/articles/{idArticle}" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Rechercher un article par ID",
        "description" : "Cette methode permet de chercher un article par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "L'article a ete trouve dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ArticleDto"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v1/categories/filter/{codeCategory}" : {
      "get" : {
        "tags" : [ "categories" ],
        "summary" : "Rechercher une categorie par CODE",
        "description" : "Cette methode permet de chercher une categorie par son CODE",
        "operationId" : "findByCode",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "codeCategory",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "L'article a ete trouve dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CategoryDto"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe dans la BDD avec le CODE fourni"
          }
        }
      }
    },
    "/gestiondestock/v1/categories/{idCategory}" : {
      "get" : {
        "tags" : [ "categories" ],
        "summary" : "Rechercher une categorie par ID",
        "description" : "Cette methode permet de chercher une categorie par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCategory",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "La categorie a ete trouve dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CategoryDto"
            }
          },
          "404" : {
            "description" : "Aucune categorie n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v1/clients/{idClient}" : {
      "get" : {
        "tags" : [ "clients" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ClientDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesclients/delete/article/{idCommande}/{idLigneCommande}" : {
      "delete" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "deleteArticle",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idLigneCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesclients/filter/{codeCommandeClient}" : {
      "get" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "findByCode",
        "parameters" : [ {
          "name" : "codeCommandeClient",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/lignesCommande/{idCommande}" : {
      "get" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "findAllLignesCommandesClientByCommandeClientId",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/LigneCommandeClientDto"
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/article/{idCommande}/{idLigneCommande}/{idArticle}" : {
      "patch" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "updateArticle",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idLigneCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/client/{idCommande}/{idClient}" : {
      "patch" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "updateClient",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/etat/{idCommande}/{etatCommande}" : {
      "patch" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "updateEtatCommande",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "etatCommande",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/quantite/{idCommande}/{idLigneCommande}/{quantite}" : {
      "patch" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "updateQuantiteCommande",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idLigneCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "quantite",
          "in" : "path",
          "required" : true,
          "type" : "number"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/{idCommandeClient}" : {
      "get" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idCommandeClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesfournisseurs/delete/article/{idCommande}/{idLigneCommande}" : {
      "delete" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "deleteArticle",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idLigneCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesfournisseurs/filter/{codeCommandeFournisseur}" : {
      "get" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "findByCode",
        "parameters" : [ {
          "name" : "codeCommandeFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesfournisseurs/update/article/{idCommande}/{idLigneCommande}/{idArticle}" : {
      "patch" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "updateArticle",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idLigneCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesfournisseurs/update/etat/{idCommande}/{etatCommande}" : {
      "patch" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "updateEtatCommande",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "etatCommande",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesfournisseurs/update/quantite/{idCommande}/{idLigneCommande}/{quantite}" : {
      "patch" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "updateQuantiteCommande",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idLigneCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "quantite",
          "in" : "path",
          "required" : true,
          "type" : "number"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesfournisseurs/{idCommandeFournisseur}" : {
      "get" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idCommandeFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/entreprises/{idEntreprise}" : {
      "get" : {
        "tags" : [ "entreprises" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idEntreprise",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/EntrepriseDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/fournisseurs/{idFournisseur}" : {
      "get" : {
        "tags" : [ "fournisseur" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/utilisateurs/{idUtilisateur}" : {
      "get" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/ventes/{codeVente}" : {
      "get" : {
        "tags" : [ "ventes" ],
        "operationId" : "findByCode",
        "parameters" : [ {
          "name" : "codeVente",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VentesDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/ventes/{idVente}" : {
      "get" : {
        "tags" : [ "ventes" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idVente",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VentesDto"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Adresse" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "AdresseDto" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "Article" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeArticle" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "photo" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "category" : {
          "$ref" : "#/definitions/Category"
        },
        "ligneVentes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneVente"
          }
        },
        "ligneCommandeClients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClient"
          }
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseur"
          }
        },
        "mvtStks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/MvtStk"
          }
        }
      }
    },
    "ArticleDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "codeArticle" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "photo" : {
          "type" : "string"
        },
        "category" : {
          "$ref" : "#/definitions/CategoryDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "Category" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "articles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Article"
          }
        }
      }
    },
    "CategoryDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "ClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "CommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "client" : {
          "$ref" : "#/definitions/ClientDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneCommandeClients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClientDto"
          }
        },
        "commandeLivree" : {
          "type" : "boolean"
        }
      }
    },
    "CommandeFournisseur" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "fournisseur" : {
          "$ref" : "#/definitions/Fournisseur"
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseur"
          }
        }
      }
    },
    "CommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "fournisseur" : {
          "$ref" : "#/definitions/FournisseurDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseurDto"
          }
        },
        "commandeLivree" : {
          "type" : "boolean"
        }
      }
    },
    "EntrepriseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "codeFiscal" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "steWeb" : {
          "type" : "string"
        }
      }
    },
    "Fournisseur" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/Adresse"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "commandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CommandeFournisseur"
          }
        }
      }
    },
    "FournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeClient" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "commandeClient" : {
          "$ref" : "#/definitions/CommandeClient"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeFournisseur" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "commandeFournisseur" : {
          "$ref" : "#/definitions/CommandeFournisseur"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "commandeFournisseur" : {
          "$ref" : "#/definitions/CommandeFournisseur"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneVente" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "vente" : {
          "$ref" : "#/definitions/Ventes"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneVenteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "vente" : {
          "$ref" : "#/definitions/VentesDto"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "RolesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "roleName" : {
          "type" : "string"
        }
      }
    },
    "UtilisateurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "dateDeNaissance" : {
          "type" : "integer",
          "format" : "int64"
        },
        "moteDePasse" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "entreprise" : {
          "$ref" : "#/definitions/EntrepriseDto"
        },
        "roles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/RolesDto"
          }
        }
      }
    },
...
    "VentesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateVente" : {
          "type" : "integer",
          "format" : "int64"
        },
        "commentaire" : {
          "type" : "string"
        },
        "ligneVentes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneVenteDto"
          }
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    }
  }
}